minor update to the rollback mihration of pre_order

+++ database/migrations/2022_08_15_091330_add_pre_order_to_products_table.php
@@ -26,7 +26,9 @@ return new class extends Migration
     public function down()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('pre_order');
+            if(Schema::hasColumn('products','pre_order')){
+                $table->dropColumn('pre_order');
+            }
         });
     }
 };
